Move tx receipt logic to separate module
