+++ core/node/api_server/src/web3/state.rs
@@ -41,12 +41,12 @@ use super::{
     backend_jsonrpsee::MethodTracer,
     mempool_cache::MempoolCache,
     metrics::{FilterType, FILTER_METRICS},
+    receipts::AccountTypesCache,
     TypedFilter,
 };
 use crate::{
     execution_sandbox::{BlockArgs, BlockArgsError, BlockStartInfo},
     tx_sender::{tx_sink::TxSink, TxSender},
-    utils::AccountTypesCache,
     web3::metrics::FilterMetrics,
 };
